internal: move getHipster to project-name
